Add cmake building system (#3)
* removed msvc project files

* updated gitignore

* removed vscode folders

* removed imgui and glm folders

* added .clang-format

* [cmake] added cmakelists

* conditional compiling for omp

* [core] fix includes

* [core] fix type

* [core] use unsigned int instead of uint

* [controller] fixes

* [gui] fix includes

* [main] fix include

* [ci] added ci

* [controller] explicit unsigned int

* [ci] add verbose to macos

* [cmake] moved required standard up

* [core] add std::

* [controller] explicit unsigned int

* [doc] add readme

* [doc] add build instructions

* [ci] disable windows for the moment

* moved imgui.ini to src and remove build dir

the file is copied inside the build directory

* fix: warning

---------

+++ src/raytracing/controller/ImageWrapper.cpp
@@ -2,12 +2,12 @@
 #include <stdio.h>
 #include <iostream>
 
-uint Raytracing::ImageWrapper::getWidth() const
+unsigned int Raytracing::ImageWrapper::getWidth() const
 {
     return width;
 }
 
-uint Raytracing::ImageWrapper::getHeight() const
+unsigned int Raytracing::ImageWrapper::getHeight() const
 {
     return height;
 }
@@ -29,7 +29,7 @@ void Raytracing::ImageWrapper::setData(uint32_t *newData)
 {
     imageData = newData;
     glBindTexture(GL_TEXTURE_2D, textureId);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, newData);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, (int) width, (int) height, 0, GL_RGBA, GL_UNSIGNED_BYTE, newData);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 }
