+++ src/raytracing/core/Scene.cpp
@@ -23,9 +23,9 @@ void Raytracing::Scene::addMaterial(const Raytracing::Material& material)
 void Raytracing::Scene::addRandomSphere()
 {
     Sphere sphere;
-    sphere.radius = (float) rand() / RAND_MAX * 2.8f + 0.2f;
-    sphere.center = glm::vec3(rand() / RAND_MAX * 10.f, rand() / RAND_MAX * 10.f, rand() / RAND_MAX * 10.f);
-    sphere.materialIndex = rand() / RAND_MAX * materialList.size();
+    sphere.radius = (float) rand() / (float)RAND_MAX * 2.8f + 0.2f;
+    sphere.center = glm::vec3((float)rand() / (float)RAND_MAX * 10.f, (float)rand() / (float) RAND_MAX * 10.f, (float) rand() / (float) RAND_MAX * 10.f);
+    sphere.materialIndex = (uint32_t) ((float)rand() / (float)RAND_MAX * (float) materialList.size());
     // create a green material if no one exist.
     if (materialList.size() == 0)
     {
